refactor: split event_parser.go by cloud provider (fixes #82)
Decompose the 987-line event_parser.go into provider-specific files:
- event_parser.go: shared types, interfaces, main dispatcher
- event_parser_aws.go: AWS CloudTrail event parsing
- event_parser_gcp.go: GCP audit log parsing
- event_parser_azure.go: Azure activity log parsing

Also split event_parser_test.go into matching test files.

No functional changes — pure structural refactoring.

Closes #82
