+++ pkg/falco/event_parser_test.go
@@ -48,64 +48,59 @@ func TestIsRelevantEvent(t *testing.T) {
 
 		// Lambda Events
 		{"Lambda UpdateFunctionConfiguration", "UpdateFunctionConfiguration", true},
-		{"Lambda CreateFunction", "CreateFunction", true},
-		{"Lambda Irrelevant", "ListFunctions", false},
+		{"Lambda UpdateFunctionCode", "UpdateFunctionCode", true},
+		{"Lambda AddPermission", "AddPermission", true},
+		{"Lambda RemovePermission", "RemovePermission", true},
+		{"Lambda Irrelevant", "GetFunction", false},
 
-		// ECS Events - Services
+		// VPC Events
+		{"VPC AuthorizeSecurityGroupIngress", "AuthorizeSecurityGroupIngress", true},
+		{"VPC AuthorizeSecurityGroupEgress", "AuthorizeSecurityGroupEgress", true},
+		{"VPC RevokeSecurityGroupIngress", "RevokeSecurityGroupIngress", true},
+		{"VPC RevokeSecurityGroupEgress", "RevokeSecurityGroupEgress", true},
+		{"VPC CreateSecurityGroup", "CreateSecurityGroup", true},
+		{"VPC DeleteSecurityGroup", "DeleteSecurityGroup", true},
+		{"VPC ModifySecurityGroupRules", "ModifySecurityGroupRules", true},
+		{"VPC CreateVpc", "CreateVpc", true},
+		{"VPC DeleteVpc", "DeleteVpc", true},
+		{"VPC Irrelevant", "DescribeSecurityGroups", false},
+
+		// KMS Events
+		{"KMS ScheduleKeyDeletion", "ScheduleKeyDeletion", true},
+		{"KMS PutKeyPolicy", "PutKeyPolicy", true},
+		{"KMS CreateKey", "CreateKey", true},
+		{"KMS Irrelevant", "DescribeKey", false},
+
+		// CloudTrail Events
+		{"CloudTrail CreateTrail", "CreateTrail", true},
+		{"CloudTrail DeleteTrail", "DeleteTrail", true},
+		{"CloudTrail UpdateTrail", "UpdateTrail", true},
+		{"CloudTrail StartLogging", "StartLogging", true},
+		{"CloudTrail StopLogging", "StopLogging", true},
+
+		// API Gateway Events
+		{"API Gateway CreateRestApi", "CreateRestApi", true},
+		{"API Gateway DeleteRestApi", "DeleteRestApi", true},
+		{"API Gateway UpdateRestApi", "UpdateRestApi", true},
+
+		// ECS Events
 		{"ECS CreateService", "CreateService", true},
 		{"ECS UpdateService", "UpdateService", true},
 		{"ECS DeleteService", "DeleteService", true},
 
-		// ECS Events - Task Definitions
-		{"ECS RegisterTaskDefinition", "RegisterTaskDefinition", true},
-		{"ECS DeregisterTaskDefinition", "DeregisterTaskDefinition", true},
+		// SNS Events
+		{"SNS CreateTopic", "CreateTopic", true},
+		{"SNS DeleteTopic", "DeleteTopic", true},
 
-		// ECS Events - Clusters
-		{"ECS UpdateCluster", "UpdateCluster", true},
-		{"ECS UpdateClusterSettings", "UpdateClusterSettings", true},
-		{"ECS PutClusterCapacityProviders", "PutClusterCapacityProviders", true},
-		{"ECS UpdateContainerInstancesState", "UpdateContainerInstancesState", true},
-
-		// ECS Events - Capacity Providers
-		{"ECS CreateCapacityProvider", "CreateCapacityProvider", true},
-		{"ECS UpdateCapacityProvider", "UpdateCapacityProvider", true},
-		{"ECS DeleteCapacityProvider", "DeleteCapacityProvider", true},
-
-		// ECS Irrelevant
-		{"ECS Irrelevant", "DescribeServices", false},
-
-		// EKS Events - Clusters
-		{"EKS CreateCluster", "CreateCluster", true},
-		{"EKS DeleteCluster", "DeleteCluster", true},
-		{"EKS UpdateClusterConfig", "UpdateClusterConfig", true},
-		{"EKS UpdateClusterVersion", "UpdateClusterVersion", true},
-
-		// EKS Events - Node Groups
-		{"EKS CreateNodegroup", "CreateNodegroup", true},
-		{"EKS DeleteNodegroup", "DeleteNodegroup", true},
-		{"EKS UpdateNodegroupConfig", "UpdateNodegroupConfig", true},
-		{"EKS UpdateNodegroupVersion", "UpdateNodegroupVersion", true},
-
-		// EKS Events - Addons
-		{"EKS CreateAddon", "CreateAddon", true},
-		{"EKS DeleteAddon", "DeleteAddon", true},
-		{"EKS UpdateAddon", "UpdateAddon", true},
-
-		// EKS Events - Fargate Profiles
-		{"EKS CreateFargateProfile", "CreateFargateProfile", true},
-
-		// EKS Irrelevant
-		{"EKS Irrelevant", "DescribeCluster", false},
-
-		// Completely irrelevant
-		{"Unknown Event", "SomeRandomEvent", false},
-		{"Empty Event", "", false},
+		// SQS Events
+		{"SQS CreateQueue", "CreateQueue", true},
+		{"SQS DeleteQueue", "DeleteQueue", true},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			got := sub.isRelevantEvent(tt.eventName)
-			assert.Equal(t, tt.want, got, "Event: %s", tt.eventName)
+			assert.Equal(t, tt.want, got, "isRelevantEvent(%q) = %v, want %v", tt.eventName, got, tt.want)
 		})
 	}
 }
@@ -119,282 +114,145 @@ func TestExtractResourceID(t *testing.T) {
 		fields    map[string]string
 		want      string
 	}{
+		// EC2 Events
 		{
-			name:      "EC2 Instance",
+			name:      "ModifyInstanceAttribute with instance ID",
 			eventName: "ModifyInstanceAttribute",
 			fields: map[string]string{
 				"ct.request.instanceid": "i-1234567890abcdef0",
 			},
 			want: "i-1234567890abcdef0",
 		},
+
+		// S3 Events
 		{
-			name:      "EBS Volume",
-			eventName: "ModifyVolume",
-			fields: map[string]string{
-				"ct.request.volumeid": "vol-123456",
-			},
-			want: "vol-123456",
-		},
-		{
-			name:      "S3 Bucket",
+			name:      "PutBucketPolicy with bucket",
 			eventName: "PutBucketPolicy",
 			fields: map[string]string{
 				"ct.request.bucket": "my-bucket",
 			},
 			want: "my-bucket",
 		},
+
+		// KMS Events
 		{
-			name:      "IAM Role",
-			eventName: "PutRolePolicy",
+			name:      "ScheduleKeyDeletion with key ID",
+			eventName: "ScheduleKeyDeletion",
 			fields: map[string]string{
-				"ct.request.rolename": "my-role",
+				"ct.request.keyid": "arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012",
 			},
-			want: "my-role",
+			want: "arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012",
+		},
+
+		// RDS Events
+		{
+			name:      "CreateDBInstance with DB instance identifier",
+			eventName: "CreateDBInstance",
+			fields: map[string]string{
+				"ct.response.dbinstanceidentifier": "my-db-instance",
+			},
+			want: "my-db-instance",
 		},
 		{
-			name:      "IAM User",
-			eventName: "CreateUser",
+			name:      "DeleteDBInstance with DB instance identifier",
+			eventName: "DeleteDBInstance",
 			fields: map[string]string{
-				"ct.request.username": "john-doe",
+				"ct.request.dbinstanceidentifier": "my-db-instance",
 			},
-			want: "john-doe",
+			want: "my-db-instance",
+		},
+
+		// VPC Events - Security Groups
+		{
+			name:      "AuthorizeSecurityGroupIngress with group ID",
+			eventName: "AuthorizeSecurityGroupIngress",
+			fields: map[string]string{
+				"ct.request.groupid": "sg-1234567890abcdef0",
+			},
+			want: "sg-1234567890abcdef0",
 		},
 		{
-			name:      "IAM Policy by ARN",
-			eventName: "CreatePolicyVersion",
+			name:      "AuthorizeSecurityGroupIngress with group name fallback",
+			eventName: "AuthorizeSecurityGroupIngress",
 			fields: map[string]string{
-				"ct.request.policyarn": "arn:aws:iam::123456789012:policy/MyPolicy",
+				"ct.request.groupname": "my-security-group",
 			},
-			want: "arn:aws:iam::123456789012:policy/MyPolicy",
+			want: "my-security-group",
+		},
+
+		// VPC Events - VPC Endpoints
+		{
+			name:      "CreateVpcEndpoint with VPC endpoint ID in response",
+			eventName: "CreateVpcEndpoint",
+			fields: map[string]string{
+				"ct.response.vpcendpointid": "vpce-12345678",
+			},
+			want: "vpce-12345678",
 		},
 		{
-			name:      "Lambda Function",
+			name:      "DeleteVpcEndpoint with VPC endpoint ID in request",
+			eventName: "DeleteVpcEndpoint",
+			fields: map[string]string{
+				"ct.request.vpcendpointid": "vpce-12345678",
+			},
+			want: "vpce-12345678",
+		},
+
+		// Lambda Events
+		{
+			name:      "UpdateFunctionConfiguration with function name",
 			eventName: "UpdateFunctionConfiguration",
 			fields: map[string]string{
 				"ct.request.functionname": "my-function",
 			},
 			want: "my-function",
 		},
-		// ECS - Services
+
+		// API Gateway - REST API
 		{
-			name:      "ECS CreateService",
-			eventName: "CreateService",
+			name:      "CreateRestApi with API ID in response",
+			eventName: "CreateRestApi",
 			fields: map[string]string{
-				"ct.response.service.servicearn": "arn:aws:ecs:us-east-1:123456789012:service/my-cluster/my-service",
+				"ct.response.id": "abcdef1234",
 			},
-			want: "arn:aws:ecs:us-east-1:123456789012:service/my-cluster/my-service",
+			want: "abcdef1234",
 		},
+
+		// CloudTrail Events
 		{
-			name:      "ECS UpdateService",
-			eventName: "UpdateService",
+			name:      "CreateTrail with trail ARN in response",
+			eventName: "CreateTrail",
 			fields: map[string]string{
-				"ct.request.service": "arn:aws:ecs:us-east-1:123456789012:service/my-cluster/my-service",
+				"ct.response.trailarn": "arn:aws:cloudtrail:us-east-1:123456789012:trail/my-trail",
 			},
-			want: "arn:aws:ecs:us-east-1:123456789012:service/my-cluster/my-service",
+			want: "arn:aws:cloudtrail:us-east-1:123456789012:trail/my-trail",
 		},
+
+		// Test fallback to default fields
 		{
-			name:      "ECS DeleteService",
-			eventName: "DeleteService",
-			fields: map[string]string{
-				"ct.request.service": "my-service",
-			},
-			want: "my-service",
-		},
-		// ECS - Task Definitions
-		{
-			name:      "ECS RegisterTaskDefinition",
-			eventName: "RegisterTaskDefinition",
-			fields: map[string]string{
-				"ct.response.taskdefinition.taskdefinitionarn": "arn:aws:ecs:us-east-1:123456789012:task-definition/my-task:1",
-			},
-			want: "arn:aws:ecs:us-east-1:123456789012:task-definition/my-task:1",
-		},
-		{
-			name:      "ECS DeregisterTaskDefinition",
-			eventName: "DeregisterTaskDefinition",
-			fields: map[string]string{
-				"ct.request.taskdefinition": "my-task:1",
-			},
-			want: "my-task:1",
-		},
-		// ECS - Clusters
-		{
-			name:      "ECS UpdateCluster",
-			eventName: "UpdateCluster",
-			fields: map[string]string{
-				"ct.request.cluster": "my-cluster",
-			},
-			want: "my-cluster",
-		},
-		{
-			name:      "ECS UpdateClusterSettings",
-			eventName: "UpdateClusterSettings",
-			fields: map[string]string{
-				"ct.request.cluster": "arn:aws:ecs:us-east-1:123456789012:cluster/my-cluster",
-			},
-			want: "arn:aws:ecs:us-east-1:123456789012:cluster/my-cluster",
-		},
-		{
-			name:      "ECS PutClusterCapacityProviders",
-			eventName: "PutClusterCapacityProviders",
-			fields: map[string]string{
-				"ct.request.cluster": "my-cluster",
-			},
-			want: "my-cluster",
-		},
-		{
-			name:      "ECS UpdateContainerInstancesState",
-			eventName: "UpdateContainerInstancesState",
-			fields: map[string]string{
-				"ct.request.containerinstances.0": "arn:aws:ecs:us-east-1:123456789012:container-instance/abc123",
-			},
-			want: "arn:aws:ecs:us-east-1:123456789012:container-instance/abc123",
-		},
-		// ECS - Capacity Providers
-		{
-			name:      "ECS CreateCapacityProvider",
-			eventName: "CreateCapacityProvider",
-			fields: map[string]string{
-				"ct.response.capacityprovider.capacityproviderarn": "arn:aws:ecs:us-east-1:123456789012:capacity-provider/my-provider",
-			},
-			want: "arn:aws:ecs:us-east-1:123456789012:capacity-provider/my-provider",
-		},
-		{
-			name:      "ECS UpdateCapacityProvider",
-			eventName: "UpdateCapacityProvider",
-			fields: map[string]string{
-				"ct.request.name": "my-provider",
-			},
-			want: "my-provider",
-		},
-		{
-			name:      "ECS DeleteCapacityProvider",
-			eventName: "DeleteCapacityProvider",
-			fields: map[string]string{
-				"ct.request.capacityprovider": "my-provider",
-			},
-			want: "my-provider",
-		},
-		// EKS - Clusters
-		{
-			name:      "EKS CreateCluster",
-			eventName: "CreateCluster",
-			fields: map[string]string{
-				"ct.response.cluster.name": "my-eks-cluster",
-			},
-			want: "my-eks-cluster",
-		},
-		{
-			name:      "EKS DeleteCluster",
-			eventName: "DeleteCluster",
-			fields: map[string]string{
-				"ct.request.name": "my-eks-cluster",
-			},
-			want: "my-eks-cluster",
-		},
-		{
-			name:      "EKS UpdateClusterConfig",
-			eventName: "UpdateClusterConfig",
-			fields: map[string]string{
-				"ct.request.name": "my-eks-cluster",
-			},
-			want: "my-eks-cluster",
-		},
-		{
-			name:      "EKS UpdateClusterVersion",
-			eventName: "UpdateClusterVersion",
-			fields: map[string]string{
-				"ct.request.name": "my-eks-cluster",
-			},
-			want: "my-eks-cluster",
-		},
-		// EKS - Node Groups
-		{
-			name:      "EKS CreateNodegroup",
-			eventName: "CreateNodegroup",
-			fields: map[string]string{
-				"ct.response.nodegroup.nodegroupname": "my-nodegroup",
-			},
-			want: "my-nodegroup",
-		},
-		{
-			name:      "EKS DeleteNodegroup",
-			eventName: "DeleteNodegroup",
-			fields: map[string]string{
-				"ct.request.nodegroupname": "my-nodegroup",
-			},
-			want: "my-nodegroup",
-		},
-		{
-			name:      "EKS UpdateNodegroupConfig",
-			eventName: "UpdateNodegroupConfig",
-			fields: map[string]string{
-				"ct.request.nodegroupname": "my-nodegroup",
-			},
-			want: "my-nodegroup",
-		},
-		{
-			name:      "EKS UpdateNodegroupVersion",
-			eventName: "UpdateNodegroupVersion",
-			fields: map[string]string{
-				"ct.request.nodegroupname": "my-nodegroup",
-			},
-			want: "my-nodegroup",
-		},
-		// EKS - Addons
-		{
-			name:      "EKS CreateAddon",
-			eventName: "CreateAddon",
-			fields: map[string]string{
-				"ct.response.addon.addonname": "vpc-cni",
-			},
-			want: "vpc-cni",
-		},
-		{
-			name:      "EKS DeleteAddon",
-			eventName: "DeleteAddon",
-			fields: map[string]string{
-				"ct.request.addonname": "vpc-cni",
-			},
-			want: "vpc-cni",
-		},
-		{
-			name:      "EKS UpdateAddon",
-			eventName: "UpdateAddon",
-			fields: map[string]string{
-				"ct.request.addonname": "vpc-cni",
-			},
-			want: "vpc-cni",
-		},
-		// EKS - Fargate Profiles
-		{
-			name:      "EKS CreateFargateProfile",
-			eventName: "CreateFargateProfile",
-			fields: map[string]string{
-				"ct.response.fargateprofile.fargateprofilename": "my-fargate-profile",
-			},
-			want: "my-fargate-profile",
-		},
-		{
-			name:      "Missing Resource ID",
-			eventName: "ModifyInstanceAttribute",
-			fields:    map[string]string{},
-			want:      "",
-		},
-		{
-			name:      "Unknown Event Type",
+			name:      "Unknown event type with fallback",
 			eventName: "UnknownEvent",
 			fields: map[string]string{
-				"ct.resource.id": "some-id",
+				"ct.resource.id": "unknown-resource-id",
 			},
-			want: "some-id",
+			want: "unknown-resource-id",
+		},
+
+		// Test with no matching fields (should return empty string)
+		{
+			name:      "Event with no matching fields",
+			eventName: "ModifyInstanceAttribute",
+			fields: map[string]string{
+				"ct.request.other": "other-value",
+			},
+			want: "",
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			got := sub.extractResourceID(tt.eventName, tt.fields)
-			assert.Equal(t, tt.want, got)
+			assert.Equal(t, tt.want, got, "extractResourceID(%q, ...) = %q, want %q", tt.eventName, got, tt.want)
 		})
 	}
 }
@@ -442,25 +300,60 @@ func TestParseFalcoOutput(t *testing.T) {
 			wantNil: true,
 		},
 		{
-			name: "Missing ct.name",
+			name: "Valid S3 PutBucketPolicy Event",
 			response: &outputs.Response{
 				Source:   "aws_cloudtrail",
 				Rule:     "AWS API Call",
 				Priority: schema.Priority_WARNING,
 				OutputFields: map[string]string{
-					"other.field": "value",
+					"ct.name":         "PutBucketPolicy",
+					"ct.request.bucket": "my-bucket",
+					"ct.src":          "s3.amazonaws.com",
+					"ct.user.type":    "IAMUser",
+					"ct.user":         "admin",
+				},
+			},
+			wantNil: false,
+		},
+		{
+			name: "Valid IAM CreateRole Event",
+			response: &outputs.Response{
+				Source:   "aws_cloudtrail",
+				Rule:     "AWS API Call",
+				Priority: schema.Priority_WARNING,
+				OutputFields: map[string]string{
+					"ct.name":           "CreateRole",
+					"ct.request.rolename": "MyRole",
+					"ct.src":            "iam.amazonaws.com",
+					"ct.user.type":      "IAMUser",
+					"ct.user":           "admin",
+				},
+			},
+			wantNil: false,
+		},
+		{
+			name: "Irrelevant AWS Event (should be filtered)",
+			response: &outputs.Response{
+				Source:   "aws_cloudtrail",
+				Rule:     "AWS API Call",
+				Priority: schema.Priority_NOTICE,
+				OutputFields: map[string]string{
+					"ct.name":   "DescribeInstances",
+					"ct.src":    "ec2.amazonaws.com",
+					"ct.user":   "admin",
 				},
 			},
 			wantNil: true,
 		},
 		{
-			name: "Irrelevant Event",
+			name: "Missing ct.name Field",
 			response: &outputs.Response{
 				Source:   "aws_cloudtrail",
 				Rule:     "AWS API Call",
-				Priority: schema.Priority_INFORMATIONAL,
+				Priority: schema.Priority_WARNING,
 				OutputFields: map[string]string{
-					"ct.name": "DescribeInstances",
+					"ct.request.instanceid": "i-1234567890abcdef0",
+					"ct.user":               "admin",
 				},
 			},
 			wantNil: true,
@@ -473,50 +366,16 @@ func TestParseFalcoOutput(t *testing.T) {
 				Priority: schema.Priority_WARNING,
 				OutputFields: map[string]string{
 					"ct.name": "ModifyInstanceAttribute",
-					// Missing instanceid
+					"ct.user": "admin",
 				},
 			},
 			wantNil: true,
-		},
-		{
-			name: "Valid S3 PutBucketEncryption Event",
-			response: &outputs.Response{
-				Source:   "aws_cloudtrail",
-				Rule:     "AWS S3 Bucket Encryption Modified",
-				Priority: schema.Priority_WARNING,
-				OutputFields: map[string]string{
-					"ct.name":           "PutBucketEncryption",
-					"ct.request.bucket": "my-secure-bucket",
-					"ct.request.serversideencryptionconfiguration": "AES256",
-					"ct.user.type":      "IAMUser",
-					"ct.user.accountid": "123456789012",
-					"ct.user":           "security-admin",
-				},
-			},
-			wantNil: false,
-		},
-		{
-			name: "Valid IAM CreateRole Event",
-			response: &outputs.Response{
-				Source:   "aws_cloudtrail",
-				Rule:     "AWS IAM Role Created",
-				Priority: schema.Priority_NOTICE,
-				OutputFields: map[string]string{
-					"ct.name":                             "CreateRole",
-					"ct.request.rolename":                 "lambda-execution-role",
-					"ct.request.assumerolepolicydocument": `{"Version":"2012-10-17","Statement":[]}`,
-					"ct.user.type":                        "IAMUser",
-					"ct.user":                             "iam-admin",
-				},
-			},
-			wantNil: false,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			got := sub.parseFalcoOutput(tt.response)
-
 			if tt.wantNil {
 				assert.Nil(t, got, "Expected nil event")
 			} else {
@@ -529,531 +388,6 @@ func TestParseFalcoOutput(t *testing.T) {
 					assert.NotEmpty(t, got.ResourceID)
 				}
 			}
-		})
-	}
-}
-
-// ============================================================================
-// AWS Parser - Edge Case Tests (Comprehensive Coverage)
-// ============================================================================
-
-func TestAWSParser_Parse_NilResponse(t *testing.T) {
-	sub := &Subscriber{}
-	event := sub.parseFalcoOutput(nil)
-	assert.Nil(t, event, "Should handle nil response gracefully")
-}
-
-func TestAWSParser_Parse_MalformedJSON(t *testing.T) {
-	sub := &Subscriber{}
-	res := &outputs.Response{
-		Source: "aws_cloudtrail",
-		OutputFields: map[string]string{
-			"ct.name":                 "ModifyInstanceAttribute",
-			"ct.request.instanceid":   "i-1234567890abcdef0",
-			"ct.request.instancetype": `{"invalid": MALFORMED_JSON}`, // Malformed JSON
-			"ct.user.type":            "IAMUser",
-			"ct.user":                 "admin",
-		},
-	}
-	event := sub.parseFalcoOutput(res)
-	assert.NotNil(t, event)
-	assert.Equal(t, "aws", event.Provider)
-	assert.Equal(t, "ModifyInstanceAttribute", event.EventName)
-}
-
-func TestAWSParser_Parse_VeryLongARN(t *testing.T) {
-	sub := &Subscriber{}
-	// ARN with 256+ characters
-	longARN := "arn:aws:iam::123456789012:user/" + string(make([]byte, 256))
-	res := &outputs.Response{
-		Source: "aws_cloudtrail",
-		OutputFields: map[string]string{
-			"ct.name":             "CreateUser",
-			"ct.request.username": "test-user-with-very-long-arn",
-			"ct.user.arn":         longARN,
-			"ct.user.type":        "IAMUser",
-			"ct.user":             "admin",
-		},
-	}
-	event := sub.parseFalcoOutput(res)
-	assert.NotNil(t, event)
-	assert.Equal(t, "test-user-with-very-long-arn", event.ResourceID)
-	assert.Equal(t, longARN, event.UserIdentity.ARN)
-}
-
-func TestAWSParser_Parse_SpecialCharactersInResourceID(t *testing.T) {
-	tests := []struct {
-		name       string
-		resourceID string
-	}{
-		{"Hyphens", "my-bucket-name-with-hyphens"},
-		{"Dots", "my.bucket.name.with.dots"},
-		{"Underscores", "my_bucket_name_with_underscores"},
-		{"Mixed", "my-bucket.name_with-all.chars"},
-		{"Numbers", "my-bucket-123456789"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sub := &Subscriber{}
-			res := &outputs.Response{
-				Source: "aws_cloudtrail",
-				OutputFields: map[string]string{
-					"ct.name":           "PutBucketEncryption",
-					"ct.request.bucket": tt.resourceID,
-					"ct.user.type":      "IAMUser",
-					"ct.user":           "admin",
-				},
-			}
-			event := sub.parseFalcoOutput(res)
-			assert.NotNil(t, event)
-			assert.Equal(t, tt.resourceID, event.ResourceID)
-		})
-	}
-}
-
-func TestAWSParser_Parse_UnicodeCharacters(t *testing.T) {
-	sub := &Subscriber{}
-	res := &outputs.Response{
-		Source: "aws_cloudtrail",
-		OutputFields: map[string]string{
-			"ct.name":             "CreateRole",
-			"ct.request.rolename": "test-role",
-			"ct.user":             "ユーザー名", // Japanese
-			"ct.user.type":        "IAMUser",
-			"ct.user.arn":         "arn:aws:iam::123456789012:user/名前", // Japanese in ARN
-			"ct.user.accountid":   "123456789012",
-		},
-	}
-	event := sub.parseFalcoOutput(res)
-	assert.NotNil(t, event)
-	assert.Contains(t, event.UserIdentity.UserName, "ユーザー")
-	assert.Contains(t, event.UserIdentity.ARN, "名前")
-}
-
-func TestAWSParser_Parse_EmptyFields(t *testing.T) {
-	tests := []struct {
-		name      string
-		fields    map[string]string
-		expectNil bool
-	}{
-		{
-			"Empty event name",
-			map[string]string{
-				"ct.name":               "",
-				"ct.request.instanceid": "i-123456",
-			},
-			true,
-		},
-		{
-			"Empty resource ID",
-			map[string]string{
-				"ct.name":               "ModifyInstanceAttribute",
-				"ct.request.instanceid": "",
-			},
-			true,
-		},
-		{
-			"Empty user identity fields",
-			map[string]string{
-				"ct.name":               "ModifyInstanceAttribute",
-				"ct.request.instanceid": "i-123456",
-				"ct.user":               "",
-				"ct.user.type":          "",
-				"ct.user.arn":           "",
-			},
-			false, // Should still parse, user identity is optional
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sub := &Subscriber{}
-			res := &outputs.Response{
-				Source:       "aws_cloudtrail",
-				OutputFields: tt.fields,
-			}
-			event := sub.parseFalcoOutput(res)
-			if tt.expectNil {
-				assert.Nil(t, event)
-			} else {
-				assert.NotNil(t, event)
-			}
-		})
-	}
-}
-
-func TestAWSParser_Parse_MultipleConsecutiveSlashesInARN(t *testing.T) {
-	sub := &Subscriber{}
-	res := &outputs.Response{
-		Source: "aws_cloudtrail",
-		OutputFields: map[string]string{
-			"ct.name":             "CreateRole",
-			"ct.request.rolename": "my-role",
-			"ct.user.arn":         "arn:aws:iam::123456789012:user//admin", // Double slash
-			"ct.user.type":        "IAMUser",
-			"ct.user":             "admin",
-		},
-	}
-	event := sub.parseFalcoOutput(res)
-	assert.NotNil(t, event)
-	assert.Equal(t, "my-role", event.ResourceID)
-	assert.Contains(t, event.UserIdentity.ARN, "//")
-}
-
-func TestAWSParser_Parse_AllAWSServiceTypes(t *testing.T) {
-	tests := []struct {
-		eventName    string
-		resourceType string
-	}{
-		// EC2
-		{"ModifyInstanceAttribute", "aws_instance"},
-		{"ModifyVolume", "aws_ebs_volume"},
-
-		// VPC
-		{"CreateSecurityGroup", "aws_security_group"},
-		{"CreateVpc", "aws_vpc"},
-		{"CreateSubnet", "aws_subnet"},
-
-		// ELB
-		{"CreateLoadBalancer", "aws_lb"},
-		{"CreateTargetGroup", "aws_lb_target_group"},
-
-		// KMS
-		{"CreateKey", "aws_kms_key"},
-		{"CreateAlias", "aws_kms_alias"},
-
-		// DynamoDB
-		{"CreateTable", "aws_dynamodb_table"},
-
-		// IAM
-		{"CreateRole", "aws_iam_role"},
-		{"CreateUser", "aws_iam_user"},
-		{"CreatePolicy", "aws_iam_policy"},
-
-		// S3
-		{"PutBucketEncryption", "aws_s3_bucket"},
-		{"PutBucketPolicy", "aws_s3_bucket_policy"},
-
-		// RDS
-		{"CreateDBInstance", "aws_db_instance"},
-		{"CreateDBCluster", "aws_rds_cluster"},
-
-		// Lambda
-		{"UpdateFunctionConfiguration", "aws_lambda_function"},
-		{"AddPermission", "aws_lambda_permission"},
-
-		// API Gateway
-		{"CreateRestApi", "aws_api_gateway_rest_api"},
-		{"CreateApi", "aws_apigatewayv2_api"},
-
-		// CloudWatch
-		{"PutMetricAlarm", "aws_cloudwatch_metric_alarm"},
-		{"CreateLogGroup", "aws_cloudwatch_log_group"},
-
-		// SNS/SQS
-		{"CreateTopic", "aws_sns_topic"},
-		{"CreateQueue", "aws_sqs_queue"},
-
-		// Route53
-		{"CreateHostedZone", "aws_route53_zone"},
-
-		// ECR
-		{"CreateRepository", "aws_ecr_repository"},
-
-		// SSM
-		{"PutParameter", "aws_ssm_parameter"},
-
-		// Secrets Manager
-		{"CreateSecret", "aws_secretsmanager_secret"},
-
-		// CloudFront
-		{"CreateDistribution", "aws_cloudfront_distribution"},
-
-		// CloudTrail
-		{"CreateTrail", "aws_cloudtrail"},
-
-		// ECS
-		{"CreateService", "aws_ecs_service"},
-		{"RegisterTaskDefinition", "aws_ecs_task_definition"},
-
-		// EKS
-		{"CreateCluster", "aws_eks_cluster"},
-		{"CreateNodegroup", "aws_eks_node_group"},
-		{"CreateAddon", "aws_eks_addon"},
-
-		// ElastiCache
-		{"CreateCacheCluster", "aws_elasticache_cluster"},
-
-		// SageMaker
-		{"CreateEndpoint", "aws_sagemaker_endpoint"},
-		{"CreateNotebookInstance", "aws_sagemaker_notebook_instance"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.eventName, func(t *testing.T) {
-			sub := &Subscriber{}
-			resourceType := sub.mapEventToResourceType(tt.eventName, "")
-			assert.Equal(t, tt.resourceType, resourceType)
-		})
-	}
-}
-
-func TestAWSParser_Parse_ConcurrentAccess(t *testing.T) {
-	sub := &Subscriber{}
-	done := make(chan bool)
-
-	for i := 0; i < 10; i++ {
-		go func(id int) {
-			res := &outputs.Response{
-				Source: "aws_cloudtrail",
-				OutputFields: map[string]string{
-					"ct.name":               "ModifyInstanceAttribute",
-					"ct.request.instanceid": string(rune(id)) + "-concurrent-test",
-					"ct.user.type":          "IAMUser",
-					"ct.user":               "test-user",
-				},
-			}
-			event := sub.parseFalcoOutput(res)
-			assert.NotNil(t, event)
-			assert.Equal(t, "aws", event.Provider)
-			done <- true
-		}(i)
-	}
-
-	for i := 0; i < 10; i++ {
-		<-done
-	}
-}
-
-func TestAWSParser_UserIdentity_EdgeCases(t *testing.T) {
-	tests := []struct {
-		name     string
-		fields   map[string]string
-		validate func(t *testing.T, userIdentity interface{})
-	}{
-		{
-			"IAMUser with full details",
-			map[string]string{
-				"ct.name":             "CreateRole",
-				"ct.request.rolename": "test-role",
-				"ct.user.type":        "IAMUser",
-				"ct.user":             "admin",
-				"ct.user.arn":         "arn:aws:iam::123456789012:user/admin",
-				"ct.user.principalid": "AIDAI123456789",
-				"ct.user.accountid":   "123456789012",
-			},
-			func(t *testing.T, userIdentity interface{}) {
-				// Just basic validation - actual structure differs
-				assert.NotNil(t, userIdentity)
-			},
-		},
-		{
-			"AssumedRole",
-			map[string]string{
-				"ct.name":             "CreateRole",
-				"ct.request.rolename": "test-role",
-				"ct.user.type":        "AssumedRole",
-				"ct.user":             "assumed-role-session",
-				"ct.user.arn":         "arn:aws:sts::123456789012:assumed-role/MyRole/MySession",
-				"ct.user.principalid": "AROAI123456789:MySession",
-				"ct.user.accountid":   "123456789012",
-			},
-			func(t *testing.T, userIdentity interface{}) {
-				assert.NotNil(t, userIdentity)
-			},
-		},
-		{
-			"Root user",
-			map[string]string{
-				"ct.name":             "CreateRole",
-				"ct.request.rolename": "test-role",
-				"ct.user.type":        "Root",
-				"ct.user":             "root",
-				"ct.user.arn":         "arn:aws:iam::123456789012:root",
-				"ct.user.accountid":   "123456789012",
-			},
-			func(t *testing.T, userIdentity interface{}) {
-				assert.NotNil(t, userIdentity)
-			},
-		},
-		{
-			"Service account",
-			map[string]string{
-				"ct.name":             "CreateRole",
-				"ct.request.rolename": "test-role",
-				"ct.user.type":        "AWSService",
-				"ct.user":             "ec2.amazonaws.com",
-				"ct.user.accountid":   "123456789012",
-			},
-			func(t *testing.T, userIdentity interface{}) {
-				assert.NotNil(t, userIdentity)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sub := &Subscriber{}
-			res := &outputs.Response{
-				Source:       "aws_cloudtrail",
-				OutputFields: tt.fields,
-			}
-			event := sub.parseFalcoOutput(res)
-			assert.NotNil(t, event)
-			assert.Equal(t, "aws", event.Provider)
-			tt.validate(t, event.UserIdentity)
-		})
-	}
-}
-
-func TestAWSParser_ExtractChanges_EdgeCases(t *testing.T) {
-	tests := []struct {
-		name      string
-		eventName string
-		fields    map[string]string
-		validate  func(t *testing.T, changes map[string]interface{})
-	}{
-		{
-			"ModifyInstanceAttribute with multiple changes",
-			"ModifyInstanceAttribute",
-			map[string]string{
-				"ct.request.instanceid":            "i-123456",
-				"ct.request.instancetype":          "t3.medium",
-				"ct.request.disableapitermination": "true",
-			},
-			func(t *testing.T, changes map[string]interface{}) {
-				assert.NotNil(t, changes)
-				// Changes should be extracted
-				if val, ok := changes["instance_type"]; ok {
-					assert.Equal(t, "t3.medium", val)
-				}
-			},
-		},
-		{
-			"PutBucketEncryption",
-			"PutBucketEncryption",
-			map[string]string{
-				"ct.request.bucket":                            "my-bucket",
-				"ct.request.serversideencryptionconfiguration": "AES256",
-			},
-			func(t *testing.T, changes map[string]interface{}) {
-				assert.NotNil(t, changes)
-				if val, ok := changes["server_side_encryption_configuration"]; ok {
-					assert.Equal(t, "AES256", val)
-				}
-			},
-		},
-		{
-			"UpdateFunctionConfiguration",
-			"UpdateFunctionConfiguration",
-			map[string]string{
-				"ct.request.functionname": "my-function",
-				"ct.request.timeout":      "300",
-				"ct.request.memorysize":   "1024",
-			},
-			func(t *testing.T, changes map[string]interface{}) {
-				assert.NotNil(t, changes)
-				// At least one change should be present
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sub := &Subscriber{}
-			changes := sub.extractChanges(tt.eventName, tt.fields)
-			tt.validate(t, changes)
-		})
-	}
-}
-
-func TestAWSParser_isRelevantEvent_EdgeCases(t *testing.T) {
-	tests := []struct {
-		name      string
-		eventName string
-		want      bool
-	}{
-		{"Empty event name", "", false},
-		{"Very long event name", "ModifyInstanceAttributeWithVeryLongNameThatExceedsNormalLength", false},
-		{"Invalid characters", "Modify@Instance#Attribute", false},
-		{"Case sensitive - wrong case", "modifyinstanceattribute", false},
-		{"Spaces in event name", "Modify Instance Attribute", false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sub := &Subscriber{}
-			got := sub.isRelevantEvent(tt.eventName)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestAWSParser_extractResourceID_EdgeCases(t *testing.T) {
-	tests := []struct {
-		name      string
-		eventName string
-		fields    map[string]string
-		want      string
-	}{
-		{
-			"Missing all possible fields",
-			"ModifyInstanceAttribute",
-			map[string]string{},
-			"",
-		},
-		{
-			"Very long resource ID",
-			"CreateLoadBalancer",
-			map[string]string{
-				"ct.response.loadbalancers.0.loadbalancerarn": "arn:aws:elasticloadbalancing:us-east-1:123456789012:loadbalancer/app/" + string(make([]byte, 256)),
-			},
-			"arn:aws:elasticloadbalancing:us-east-1:123456789012:loadbalancer/app/" + string(make([]byte, 256)),
-		},
-		{
-			"Unknown event type with fallback",
-			"UnknownEvent",
-			map[string]string{
-				"ct.resource.id": "fallback-id",
-			},
-			"fallback-id",
-		},
-		{
-			"Empty string values",
-			"ModifyInstanceAttribute",
-			map[string]string{
-				"ct.request.instanceid": "",
-			},
-			"",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sub := &Subscriber{}
-			got := sub.extractResourceID(tt.eventName, tt.fields)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestAWSParser_mapEventToResourceType_UnknownEvent(t *testing.T) {
-	sub := &Subscriber{}
-
-	tests := []struct {
-		eventName string
-		expected  string
-	}{
-		{"UnknownEvent", "unknown"},
-		{"SomeRandomEvent", "unknown"},
-		{"", "unknown"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.eventName, func(t *testing.T) {
-			got := sub.mapEventToResourceType(tt.eventName, "")
-			assert.Equal(t, tt.expected, got)
 		})
 	}
 }
